% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocor_plot.R
\name{autocor_plot}
\alias{autocor_plot}
\title{Autocorrelation Plot}
\usage{
autocor_plot(
  your_SE,
  method_corr = "pearson",
  filter_by,
  filter_selection,
  plot_over,
  plot_over_display_choices = NULL,
  keep_numeric = FALSE,
  your_title = "",
  no_negatives = FALSE,
  point_size = 1,
  line_size = 1,
  text_size = 15,
  return_table = FALSE
)
}
\arguments{
\item{your_SE}{Summarized Experiment object containing clonal tracking data as created by the barcodetrackR `create_SE` function.}

\item{method_corr}{Character. One of "pearson", "spearman", or "kendall".}

\item{filter_by}{Name of metadata column to filter by e.g. Lineage}

\item{filter_selection}{The value of the filter column to display e.g. "T" (within Lineage)}

\item{plot_over}{The column of metadata that you want to be the x-axis of the plot. e.g. Month}

\item{plot_over_display_choices}{Choice(s) from the column designated in plot_over that will be used for plotting. Defaults to all.}

\item{keep_numeric}{If plot_over is numeric, whether to space the x-axis appropriately according to the numerical values.}

\item{your_title}{The title for the plot.}

\item{no_negatives}{Logical. Whether to make negative correlations = 0.}

\item{point_size}{Numeric. Size of the points.}

\item{line_size}{Numeric. Thickness of the lines.}

\item{text_size}{Numeric. Size of text in plot.}

\item{return_table}{Logical. If set to true, insetad of returning a plot, the function will return a dataframe containing the pairwise correlation values between samples.}
}
\value{
Plots pairwise correlation plot for the samples in your_SE. Or, a dataframe of the pairwise correlation values if return_table is set to TRUE.
}
\description{
Gives the pairwise correlation between each sample-sample pair in the data frame in a group-by-group manner. Considers all samples matching the provided "filter_selection" within the "filter_by" column of your metadata. For each unique value of the "plot_over" argument, plots the correlation of that sample with all other samples in that group as a line plot.
}
\examples{
autocor_plot(six_SE, plot_over = "months", filter_by = "celltype", filter_selection = "Gr", no_negatives = TRUE, keep_numeric = TRUE)
}
