% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_ggheatmap_2.R
\name{barcode_ggheatmap_2}
\alias{barcode_ggheatmap_2}
\title{barcode_ggheatmap (Barcode Heatmap using ggplot2)}
\usage{
barcode_ggheatmap_2(
  your_SE,
  assay = "logs",
  selections = list(),
  plot_labels = NULL,
  n_clones = 10,
  your_title = "",
  grid = TRUE,
  label_size = 1,
  dendro = FALSE,
  cellnote_size = 4,
  distance_method = "Euclidean",
  minkowski_power = 2,
  cellnote_assay = "stars",
  hclust_linkage = "complete",
  row_order = "hierarchical",
  clusters = 0,
  percent_scale = c(0, 2.5e-05, 0.001, 0.01, 0.1, 1)
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{assay}{Perform visualization and clustering on this assay.}

\item{selections}{A list of selections to use leveraging the metadata information (see subset_SE for the necessary formatting, e.g selections = list("month" > 1, "celltype" = c("T", "B")))}

\item{plot_labels}{Vector of x axis labels. Defaults to colnames(your_SE).}

\item{n_clones}{The top 'n' clones to plot.}

\item{your_title}{The title for the plot.}

\item{grid}{Logical. Include a grid or not in the heatmap.}

\item{label_size}{The size of the column labels.}

\item{dendro}{Logical. Whether or not to show row dendrogram when hierarchical clustering.}

\item{cellnote_size}{The numerical size of the cell note labels.}

\item{distance_method}{Character. Use summary(proxy::pr_DB) to see all possible options for distance metrics in clustering.}

\item{minkowski_power}{The power of the Minkowski distance (if minkowski is the distance method used).}

\item{hclust_linkage}{Character. One of one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{row_order}{Character; "hierarchical" to perform hierarchical clustering on the output and order in that manner, "emergence" to organize rows  by order of presence in data (from left to right), or a character vector of rows within the summarized experiment to plot.}

\item{clusters}{How many clusters to cut hierarchical tree into for display when row_order is "hierarchical".}

\item{percent_scale}{A numeric vector of length 5 through which to spread the color scale (from 0% to 100%). Defaults to c(0, 0.000025, 0.001, 0.01, 0.1, 1).}

\item{cellnote_option}{Character. One of "stars", "reads", or "percents"}
}
\value{
Displays a heatmap in the current plot window.
}
\description{
Creates a heatmap using the top 'n' rows from each column in the Summarized Experiment object, using ggplot2.
}
\examples{
barcode_ggheatmap_2(your_SE = ZH33_SE,  n_clones = 100,  grid = TRUE, label_size = 3)
}
