% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_statistics.R
\name{barcode_statistics}
\alias{barcode_statistics}
\title{Barcode Statistics}
\usage{
barcode_statistics(
  your_SE,
  sample_size,
  stat_test = "chi-squared",
  stat_option = "subsequent",
  reference_sample = NULL,
  bc_threshold = 0
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{sample_size}{A numeric vector providing the sample size of each column of the SummarizedExperiment passed to the function. This sample size describes the samples that the barcoding data is meant to approximate.}

\item{stat_test}{The statistical test to use on the constructed contingency table for each barcoe. Options are "chi-squared" and "fisher."}

\item{stat_option}{For "subsequent" statistical testing is performed on each column of data compared to the column before it. For "reference," all other columns of data are compared to a reference column.}

\item{reference_sample}{Provide the column name of the reference column if stat_option is set to "reference." Defaults to the first column in the SummarizedExperiment.}

\item{bc_threshold}{Clones must be above this proportion in at least one sample to be included in statistical testing.}
}
\value{
Displays a heatmap in the current plot window.
}
\description{
Calculates statistical testing comparing each barcode between samples.
}
\examples{
barcode_statistics(your_SE = ZH33_SE, sample_size = rep(5000,ncol(ZH33_SE)), stat_test = "chi-squared", stat_option = "subsequent", bc_threshold = 0.0001)

}
