% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{barcode_treemap}
\alias{barcode_treemap}
\title{Barcode treemap}
\usage{
barcode_treemap(your_data, column_choice = NULL, threshold = 0,
  colors_choice = c("blue", "yellow"), your_title = "")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{column_choice}{Character. The name of the column to make the treemap from.}

\item{threshold}{Numeric. Threshold for the whole data frame. Anything lower than this becomes 0.}

\item{colors_choice}{Character vector of size 2 specifying which colors to use for the treemap range.}

\item{your_title}{Character. Title for the plot.}
}
\value{
Displays a treemap (made by treemap) of the chosen column.
}
\description{
Makes a treemap of the chosen column in your data frame.
}
\examples{
barcode_treemap(your_data = zh33, column_choice = "32m_CD4.fastq", colors_choice = c("red", "green"))
}

