% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodecount.R
\name{barcodecount}
\alias{barcodecount}
\title{Barcode Counter}
\usage{
barcodecount(your_data, count = "cumulative", months = NULL,
  threshold = 0)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{count}{Character. One of "cumulative", "unique", or "new".}

\item{months}{Numeric vector. Corresponds to months of each column of your_data.}

\item{threshold}{Numeric. Universal threshold applied to your_data (anything below this threshold will be set to 0.}
}
\value{
Vector containing the barcode count of each column, named by month.
}
\description{
Counts the number of barcodes sequentially by column cumulatively, uniquely, or by first presence (new).
}
\examples{
barcodecount(your_data = zh33, count = "unique", months = c(1,2,3,4.5,6.5,9), threshold = 1000)
barcodecount(your_data = zg66, count = "new", months = c(1:10), threshold = 900)
}

