% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_clusters.R
\name{BC_clusters}
\alias{BC_clusters}
\title{BC_clusters (Barcode Clusters)}
\usage{
BC_clusters(your_data, names = colnames(your_data), n_clones = 10,
  months = 1:ncol(your_data), your_title = "", log_transform = TRUE,
  log_choice = exp(1), variable_log_min = TRUE,
  distance_method = "Euclidean", minkowski_power = 1,
  hclust_linkage = "complete", clusters = 3, percent_scale = TRUE)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{names}{Names for each column of the data frame.}

\item{n_clones}{The top 'n' clones to plot.}

\item{your_title}{The title for the plot.}

\item{log_transform}{Logical. Log transform data before clustering and plotting.}

\item{log_choice}{Data is log transformed with this log.}

\item{variable_log_min}{Logical. TRUE to make log(0) = minimum nonzero - 1. FALSE to make log(0) = log(100/4000000) - 1.}

\item{distance_method}{Character. Use summary(proxy::pr_DB) to see all options.}

\item{minkowski_power}{The power of the Minkowski distance (if minkowski is used).}

\item{hclust_linkage}{Character. One of one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{clusters}{How many clusters to cut hierarchical tree into for display when row_order is "hierarchical".}

\item{percent_scale}{Logical. Whether to display plot with percent scale rather than log scale.}
}
\value{
Displays a line plot in the current plot window.
}
\description{
Using the same methodlogy as BCheatmap, takes the top 'n' rows from each column, clusters them, and plots
their average values over each sample (over time).
}
\examples{
BCheatmap(your_data = zh33, names = colnames(zh33), n_clones = 10,
      your_title = "First Time Point", grid = TRUE, columnLabels = 3)
BCheatmap(your_data = zh33, n_clones = 10, printtable = TRUE)
}

