% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCLH.R
\name{BCLH}
\alias{BCLH}
\title{Barcode Log Read Fraction of Hematopoiesis}
\usage{
BCLH(your_data, names = colnames(your_data), n_clones = 10,
  your_title = "", grid = TRUE, columnLabels = 1, dendro = "none",
  writekey = FALSE, star_size = 1)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{names}{Names for each column of the data frame.}

\item{n_clones}{The top 'n' clones to plot.}

\item{your_title}{The title for the plot.}

\item{grid}{Logical. Include a grid or not in the heatmap.}

\item{columnLabels}{The size of the column labels.}

\item{dendro}{Whether or not to rearrange data with dendrogram.}

\item{writekey}{Logical. Write contributions to file or not.}
}
\value{
Displays a heatmap in the current plot window.
}
\description{
Creates a heatmap using the top 'n' rows from each column. Data undergoes a natural log transformation before being plotted. Optionally, will write a percentage key to the current directory.
}
\examples{
BCLH(your_data = zh33, names = colnames(zh33), n_clones = 10,
      your_title = "First Time Point", grid = TRUE, columnLabels = 3,
         writekey = FALSE)

}

