% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_dot.R
\name{bias_dot}
\alias{bias_dot}
\title{Bias dot plot}
\usage{
bias_dot(
  your_SE,
  cell_var,
  cell_1,
  cell_2,
  filter_by,
  filter_selection,
  text_size = 20,
  alpha = 0.5
)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{cell_var}{The column of metadata corresponding to cell types}

\item{cell_1}{The first cell type to be compared}

\item{cell_2}{The second cell type to be compared}

\item{filter_by}{The column of metadata to filter by. e.g. Timepoint}

\item{filter_selection}{The value of filter_by you want to create histogram for e.g. 2 months}

\item{text_size}{The size of the text in the plot.}

\item{alpha}{The transparency of the dots. Lower is more translucent, 1 is completely opaque}
}
\value{
Histogram of log bias for two lineages over time.
}
\description{
Given a summarized experiment, gives dot plot of log biases for 2 cell types
}
\examples{
bias_dot(your_SE = SE, cell_var = "Cell_type", cell_1 = "B", cell_2 = "T", filter_by = "Timepoint", filter_selection = "3m")
}
