% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_histogram.R
\name{bias_histogram}
\alias{bias_histogram}
\title{Bias histogram}
\usage{
bias_histogram(
  your_SE,
  split_bias_on,
  bias_1,
  bias_2,
  split_bias_over,
  bias_over = NULL,
  remove_unique = FALSE,
  breaks = c(10, 2, 1, 0.5),
  text_size = 10,
  linesize = 0.4,
  ncols = 1,
  scale_all_y = TRUE
)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{split_bias_on}{The column in `colData(your_SE)` from which `bias_1` and `bias_2` will be chosen}

\item{bias_1}{The factor you wish to plot on the right side of the plots.}

\item{bias_2}{The factor you wish to plot on the left side of the plots.}

\item{split_bias_over}{The column in `colData(your_SE)` that you wish to observe the bias split on.}

\item{bias_over}{Choice(s) from the column designated in `split_bias_over` that will be used for plotting. Defaults to all.}

\item{remove_unique}{If set to true, only clones present in both samples will be considered.}

\item{breaks}{Numeric. The breaks specified for bins on the x-axis (how biased the clones are towards one factor or the other).}

\item{text_size}{The size of the text in the plot.}

\item{linesize}{The linewidth of the stacked bars which represent individual barcodes}

\item{ncols}{Numeric. Number of columns to plot on using plot_grid from cowplot.}

\item{scale_all_y}{Logical. Whether or not to plot all plots on the same y axis limits.}
}
\value{
Histogram of log bias for two factors over another set of factors.
}
\description{
Given a summarized experiment, gives histogram of log biases for 2 cell types
}
\examples{
bias_histogram(your_SE = SE, split_bias_on = "Lineage", bias_1 = "B", bias_2 = "T", split_bias_over = "Month", bias_over = c(1,2,4.5,12))
}
