% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_lineplot.R
\name{bias_lineplot}
\alias{bias_lineplot}
\title{Bias line plot}
\usage{
bias_lineplot(
  your_SE,
  split_bias_on,
  bias_1,
  bias_2,
  split_bias_over,
  bias_over = NULL,
  remove_unique = FALSE,
  text_size = 16,
  keep_numeric = TRUE
)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{split_bias_on}{The column of metadata corresponding to cell types}

\item{bias_1}{The first cell type to be compared. Will be on the UPPER side of the line plot}

\item{bias_2}{The second cell type to be compared. Will be on the LOWER side of the line plot}

\item{split_bias_over}{The column of metadata to plot by. If numeric, y axis will be in increasing order. If categorical, it will follow order of metadata.}

\item{bias_over}{Choice(s) from the column designated in `split_bias_over` that will be used for plotting. Defaults to all.}

\item{remove_unique}{If set to true, only clones present in both samples will be considered.}

\item{text_size}{The size of the text in the plot.}

\item{keep_numeric}{Whether to keep the numeric spacing within split_bias_over or switch to discrete x scale}
}
\value{
Bias line plot for two lineages over time.
}
\description{
Given a summarized experiment, gives ridge plots showing percent total contribution to both lineages.
}
\examples{
bias_lineplot(your_se = SE, split_bias_on = "selection_type", bias_1 = "B", bias_2 = "T", split_bias_over = "Timepoint")
}
