% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index_html.R
\name{build_index_html}
\alias{build_index_html}
\title{Build html}
\usage{
build_index_html(
  target = "vignettes/Introduction_to_barcodetrackR.Rmd",
  output = "index.html"
)
}
\description{
Build html for vignette to index.html in docs
}
