% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_plot.R
\name{circos_plot}
\alias{circos_plot}
\title{Barcode Circos plot}
\usage{
circos_plot(your_SE, weighted = FALSE, plot_label = "SAMPLENAME", alpha = 1)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{weighted}{weighted = F which is default will make links based on # of shared clones. Weighted = T will make links based on their proportion.}

\item{plot_label}{Name of colData variable to use as labels for regions. Defaults to SAMPLENAME}

\item{alpha}{Transparency of links. Default = 1 is opaque. 0 is completely transluscent}
}
\value{
Displays a circos plot in the current plot window.
}
\description{
Creates a circos plot showing each cell type as a region around a circle and shared clones between these cell types as links between the regions.
}
\examples{
circos_plot(your_SE = ZH33_SE,  plot_label = 'Cell_type')

}
