% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_bias.R
\name{clonal_bias}
\alias{clonal_bias}
\title{Clonal Bias}
\usage{
clonal_bias(your_data, text_size = 20)
}
\arguments{
\item{your_data}{Your data frame.}

\item{text_size}{The size of the text in the plot.}
}
\value{
Bias plot for two lineages over time.
}
\description{
Given a data frame organized with two alternating cell lineages, gives bias plots showing percent total contribution to both lineages.
}
\examples{
clonal_bias(datum[,c("T 1m", "B 1m", "T 3m" "B 3m")])
}
