% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_contribution.R
\name{clonal_contribution}
\alias{clonal_contribution}
\title{Bar or line Plot of Hematopoietic Contribution from top n clones}
\usage{
clonal_contribution(
  your_SE,
  graph_type = "bar",
  filter_by,
  filter_selection,
  plot_by,
  n_clones = 10,
  linesize = 0.5,
  text_size = 15
)
}
\arguments{
\item{your_SE}{A summarized experiment. With barcode data stored in the 1st assay and metadata stored in colData.}

\item{graph_type}{Choice of "bar" or "line" for how to display the clonal contribution data}

\item{filter_by}{Name of metadata column to filter by e.g. Cell_type}

\item{filter_selection}{The value of the filter column to display e.g. T cells}

\item{plot_by}{The column of metadata that you want to be the x-axis of the plot. e.g. Timepoint}

\item{n_clones}{Numeric. Number of top clones from each sample that should be displayed.}

\item{linesize}{Numeric. Thickness of the lines. OPTIONAL}

\item{text_size}{Numeric. Size of text in plot. OPTIONAL}
}
\value{
Displays a stacked area plot (made by ggplot2) of the samples' top clones.
}
\description{
Usually used for tracking a cell lineage's top clones over time.
}
\examples{
clonal_contribution(your_data = ZG66_simple_data, graph_type = "bar",filter_by = "Cell_type", filter_selection = "B", plot_by = "Timepoint", n_clones = 20)
}
