% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonaldiversity.R
\name{clonaldiversity}
\alias{clonaldiversity}
\title{Clonal Diversity}
\usage{
clonaldiversity(your_data, threshold = 2000, divindex = "shannon")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{threshold}{Numeric. Universal threshold applied to your_data (anything below this threshold will be set to 0.#'@param writekey Logical. Write contributions to file or not.}

\item{divindex}{Character. One of "shannon", "simpson", or "invsimpson".}
}
\value{
Vector of the diversities of each column.
}
\description{
Gives the diversity for each sample (column) in your data.
}
\examples{
clonaldiversity(your_data = zh33, threshold = 2000, divindex = "shannon")
}

