% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_clones.R
\name{count_clones}
\alias{count_clones}
\title{Clonal count plot}
\usage{
count_clones(
  your_SE,
  plot_over,
  plot_over_display_choices = NULL,
  group_by,
  group_by_choices = NULL,
  index_type = "count",
  point_size = 3,
  line_size = 2,
  text_size = 12,
  your_title = NULL
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{plot_over}{The column of metadata that you want to be the x-axis of the plot. e.g. timepoint}

\item{plot_over_display_choices}{Choice(s) from the column designated in plot_over that will be used for plotting. Defaults to all if left as NULL.}

\item{group_by}{The column of metadata you want to group by e.g. cell_type}

\item{group_by_choices}{Choice(s) from the column designated in group_by that will be used for plotting. Defaults to all if left as NULL.}

\item{index_type}{Character. One of "count" or "cumulative_count".}

\item{point_size}{Numeric. Size of points.}

\item{line_size}{Numeric. Size of lines.}

\item{text_size}{Numeric. Size of text in plot.}

\item{your_title}{The title for the plot.}
}
\value{
Outputs plot of a diversity measure tracked for groups over a factor.
}
\description{
A line plot that tracks a diversity measure from a selected choice or number of elements in the rows of the SummarizedExperiment object.
}
\examples{
count_clones(your_data = wu_SE, index_type = "shannon", plot_by = timepoint, group_by = cell_type)

}
