% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE}
\alias{create_SE}
\title{create_SE}
\usage{
create_SE(
  your_data = NULL,
  meta_data = NULL,
  threshold = 0,
  log_base = exp(1),
  scale_factor = 1e+06
)
}
\arguments{
\item{your_data}{A data frame. For clonal tracking data, this will be individual barcodes or lineage tracing elements in rows and samples in columns.}

\item{meta_data}{A data frame containing all meta-data. Must, at the very least, include a column called "SAMPLENAME" that contains all of the colnames within the data frame passed as  `your_data` and only those colnames.}

\item{threshold}{Optional. A numeric indicating which reads to filter out (only those rows which have higher than `threshold` proportion of reads within at least one sample will be kept as non-zero) prior to creating the SE object.}

\item{log_base}{A numeric indicating which base to use when logging the percentages.}

\item{scale_factor}{A numeric indicating what scaling factor to use in normalization. For the default value of 1 million, barcode percentages on a per sample basis will be multiplied by 1 million before log+1 normalization.}
}
\value{
Returns a SummarizedExperiment holding your clonal tracking data and the associated metadata.
}
\description{
Creates a SummarizedExperiment object from a data frame containing clonal tracking counts (`your_data`) with rows as observations and columns as samples, and the associated metadata (`meta_data`) with rows as samples and columns of information describing those samples.
}
\examples{
create_SE(your_data = exp1_clonal_counts_data, meta_data = exp1_meta_data)

}
