% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_plot.R
\name{diversity_plot}
\alias{diversity_plot}
\title{Diversity Plot}
\usage{
diversity_plot(
  your_SE,
  index_type = "shannon",
  measure = "diversity",
  plot_by,
  group_by,
  point_size = 5,
  line_size = 2
)
}
\arguments{
\item{your_SE}{A summarized experiment. With barcode data stored in the 1st assay and metadata stored in colData.}

\item{index_type}{Character. One of "herfindahl-hirschman", "shannon", "simpson", or "invsimpson".}

\item{measure}{The measure you want to plot. Can be "diversity", "count", or "evenness"}

\item{plot_by}{The column of metadata that you want to be the x-axis of the plot. e.g. timepoint}

\item{group_by}{The column of metadata you want to group by e.g. cell _type}

\item{point_size}{Numeric. Size of points. Optional}

\item{line_size}{Numeric. Size of lines. Optional}
}
\value{
Outputs plot of diversity tracked over time.
}
\description{
Plots (using ggplot2) a line plot that tracks the diversity of a lineage of cells across time points.
}
\examples{
diversity_plot(your_data = ZG66_simple_data, index_type = "shannon",
               measure = "diversity", plot_by = timepoint, group_by = cell_type)

}
