% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_bias.R
\name{dot_bias}
\alias{dot_bias}
\title{Dot Bias}
\usage{
dot_bias(your_data, text_size = 20, bias_form = "fold")
}
\arguments{
\item{your_data}{Your data frame.}

\item{text_size}{The size of the text in the plot.}

\item{bias_form}{One of "log2", "fractional", "fold" specifying how to calculate bias.}
}
\value{
Bias plot for two lineages over time.
}
\description{
Given a data frame organized with two alternating cell lineages, gives bias jittered dot plots showing percent total contribution to both lineages over time.
}
\examples{
clonal_bias(datum[,c("T 1m", "B 1m", "T 3m" "B 3m")])
}

