% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettopindices.R
\name{gettopindices}
\alias{gettopindices}
\title{Top Clones Getter}
\usage{
gettopindices(your_data, top = 5)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{top}{Numeric. How many top entries to find.}
}
\value{
Vector of indices of the top N entries of each column.
}
\description{
Given a subset of data, returns only the rows that contain the top N entries of each column.
}
\examples{
gettopindices(your_data = zh33[,c(1:5)], top = 10)

}

