% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hematoper.R
\name{hematoper}
\alias{hematoper}
\title{Hematopoietic Contribution}
\usage{
hematoper(your_data, months, n_clones = 10, linesize = 0.5,
  text_size = 15, y_limit = 100, plot_theme = "classic")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{months}{Numeric vector. Vector corresponding to the months of each sample.}

\item{n_clones}{Numeric. Number of top clones from each month that should be displayed.}

\item{linesize}{Numeric. Thickness of the lines.}

\item{text_size}{Numeric. Size of text in plot.}

\item{y_limit}{Numeric. Limit of y axis in plot.}

\item{plot_theme}{Character. One of "BW", "classic", or "original".}
}
\value{
Displays a stacked area plot (made by ggplot2) of the samples' top clones.
}
\description{
Usually used for tracking a cell lineage's top clones over time. Best to
put samples in order from left to right (i.e. "1mT", "2mT", "3mT", etc.)
}
\examples{
hematoper(your_data = zh33_Tlineage, months = c(1,2,3,4,5,6.5, 9.5, 12), y_limit = 50)
}

