% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch Barcode App}
\usage{
launchApp(x = NULL)
}
\arguments{
\item{x}{NULL}
}
\value{
Page launching the Shiny Barcode App
}
\description{
Launches the Shiny Barcode App.
}
\examples{
launchApp()
}
