% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchAppII.R
\name{launchAppII}
\alias{launchAppII}
\title{Launch Multimonkey App}
\usage{
launchAppII()
}
\value{
Page launching the Shiny Barcode App
}
\description{
Launches the Shiny Multimonkey App.
}
\examples{
launchAppII()
}

