% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_stacked.R
\name{library_stacked}
\alias{library_stacked}
\title{Library Stacked area plot}
\usage{
library_stacked(list_of_libraries, library_titles, samples, time,
  title = NULL, colors = FALSE, mode = "PROP", n = 9, Direction = 1)
}
\arguments{
\item{list_of_libraries}{A list of data frames. Data frames usually have individual barcodes in rows and samples in columns.}

\item{library_titles}{The names of each library used}

\item{samples}{A vector of the sample names you wish to include, in time order.}

\item{time}{The numeric time points associated with samples, in ascending order.}

\item{title}{The title.}

\item{colors}{The color palette of the figure. Default is from the viridis package.}

\item{mode}{The type of stacked area plot. "PROP" is a proportioned stacked area plot, displaying the proportion of each barcode at the time points. "Plain" displays the pure counts of each barcode.}

\item{n}{n is only relevant if using a non default color palette and using more clones than the size of the palette. If n_clones is greater than the size of the palette, n should be the max size of the palette.}

\item{Direction}{The direction of the color palette used. Appropriate inputs are +1 or -1.}
}
\value{
Displays stacked area plot in the current plot window.
}
\description{
Creates a stacked area plot over time for the total counts for all barcodes in a given library in your data set.
}
\examples{
library_stacked(list(lib7_jd76,lib11_jd76),c("Library 7","Library 11"),c("jd76_1m_grans.fastq","jd76_2m_grans.fastq","jd76_3m_grans.fastq"),c(1,2,3),20)
}
