% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merger}
\alias{merger}
\title{Data Frame Merger}
\usage{
merger(...)
}
\arguments{
\item{...}{A list of data frames to merge by row name.}
}
\value{
Merged data frame.
}
\description{
Merges an arbitrary number of data frames together by row name, including all entries. Hi, mom!
}
\examples{
merger(dataframe1, dataframe2, dataframe3)
}

