% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimonkey_richness.R
\name{multimonkey_richness_plot}
\alias{multimonkey_richness_plot}
\title{MultiMonkey Richness
Plots (using ggplot2) a line plot that tracks the barcode counts of cell types across time points and monkeys. Ideally should only be
used with the app, as the data set up can be quite cumbersome otherwise.}
\usage{
multimonkey_richness_plot(outfile_list, outfile_names, outfile_months_list,
  celltypes_list, threshold_list, combine = FALSE, richness_type = "unique",
  point_size = 5, line_size = 3, y_lower = 0, y_upper = 9.5)
}
\arguments{
\item{outfile_list}{A list of outfiles as data frames. See note below for organization of each data frame.}

\item{outfile_names}{A list of names that correspond to the monkey names of your outfiles.}

\item{outfile_months_list}{A list of the months for each of the corresponding data frames.}

\item{celltypes_list}{A list of the celltypes for each of the corresponding data frames.}

\item{threshold_list}{A list of the thresholds for each of the corresponding data frames.}

\item{combine}{Logical. Whether to combine the monthly barcodes.}

\item{richness_type}{Character. One of "unique", "cumulative", or "new".}

\item{point_size}{Numeric. Size of points.}

\item{line_size}{Numeric. Size of lines.}

\item{y_lower}{Numeric. Lower bound of y axis.}

\item{y_upper}{Numeric. Upper bound of y axis.}
}
\value{
Outputs plot of barcode richness tracked over time.
}
\description{
MultiMonkey Richness
Plots (using ggplot2) a line plot that tracks the barcode counts of cell types across time points and monkeys. Ideally should only be
used with the app, as the data set up can be quite cumbersome otherwise.
}
\note{
Each data frame must be organized in columns by month, and then by cell type. For example, when tracking B and T cells
over three months, your columns should be in this order: "1m T", "1m B", "2m T", "2m B", "3m T", "3m B". Then
celltypes should be c("T", "B"), and months should be c(1,2,3). Repeat this for every monkey.
}
\examples{
multimonkey_richness_plot(outfile_list = list(zh33, zg66, zh19),
                         outfile_names = list("zh33", "zg66", "zh19"),
                         celltypes_list = list(c("T", "B", "Gr"), c("T"), c("T, "B", "Gr)),
                         threshold_list = list(1000,1000,0),
                         combine = FALSE,
                         richness_type = "new")


}

