% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radartopclones.R
\name{radartopclones}
\alias{radartopclones}
\title{Radar Top Clones}
\usage{
radartopclones(
  your_data,
  columnChoice_Name = "",
  n_clones = 10,
  your_title = "",
  labelsize = 1
)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{columnChoice_Name}{Character. The name of the column to take the n clones from.}

\item{n_clones}{Numeric. The number of top clones to take from columnChoice_Name.}
}
\value{
labelsize Numeric. Size of the labels.
}
\description{
Makes a radar chart of your data frame, plotting the percentage contribution from
the top N clones from a specified column.
}
\examples{

radartopclones(your_data = zh33[,c(1:6)], columnChoice_Name = "zh331mT.fastq",
              your_title = "1m Radar Plot")
}
