% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_plotter.R
\name{richness_plot}
\alias{richness_plot}
\title{Richness Plotter}
\usage{
richness_plot(your_data, months, celltypes, thresh = 0, point_size = 5,
  line_size = 3, richness_type = "unique", y_lower = 0, y_upper = 2000)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{months}{Numeric vector. Corresponds to the months of your_data.}

\item{celltypes}{Character vector. Corresponds to the celltypes of your_data.}

\item{thresh}{Numeric. Universal threshold applied to your_data (anything below this threshold will be set to 0)}

\item{point_size}{Numeric. Size of points.}

\item{line_size}{Numeric. Size of lines.}

\item{richness_type}{Character. One of "unique", "cumulative", or "new".}

\item{y_lower}{Numeric. Lower bound of y axis.}

\item{y_upper}{Numeric. Upper bound of y axis.}
}
\value{
Outputs plot of diversity tracked over time.
}
\description{
Plots (using ggplot2) a line plot that tracks the barcode counts of cell types across time points.
}
\note{
Data must be organized in columns by month, and then by cell type. For example, when tracking B and T cells
over three months, your columns should be in this order: "1m T", "1m B", "2m T", "2m B", "3m T", "3m B". Then
celltypes should be c("T", "B"), and months should be c(1,2,3).
}
\examples{
richness_plot(your_data = zh33_first3m_T_B_NK,
               months = c(1,2,3), celltypes = c("T", "B", "NK"),
               thresh = 1000, richness_type = "cumulative")


}

