% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_plot.R
\name{ridge_plot}
\alias{ridge_plot}
\title{Ridge plot}
\usage{
ridge_plot(
  your_SE,
  split_bias_on,
  bias_1,
  bias_2,
  split_bias_over,
  bias_over = NULL,
  remove_unique = FALSE,
  weighted = F,
  text_size = 16,
  scale = 1
)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{split_bias_on}{The column of metadata corresponding to cell types}

\item{bias_1}{The first cell type to be compared. Will be on the RIGHT side of the ridge plot}

\item{bias_2}{The second cell type to be compared. Will be on the LEFT side of the ridge plot}

\item{split_bias_over}{The column of metadata to plot by. If numeric, y axis will be in increasing order. If categorical, it will follow order of metadata.}

\item{bias_over}{Choice(s) from the column designated in `split_bias_over` that will be used for plotting. Defaults to all.}

\item{remove_unique}{If set to true, only clones present in both samples will be considered.}

\item{weighted}{If true, the density estimation will be weighted by the overall contribution of each barcode}

\item{text_size}{The size of the text in the plot.}

\item{scale}{Sets the overlap between ridges. Larger values give more overlap.}
}
\value{
Bias plot for two lineages over time.
}
\description{
Given a summarized experiment, gives ridge plots showing percent total contribution to both lineages.
}
\examples{
ridge_plot(your_se = SE, split_bias_on = "selection_type", bias_1 = "B", bias_2 = "T", split_bias_over = "Timepoint")
}
