% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_plot_weighted.R
\name{ridge_plot_weighted}
\alias{ridge_plot_weighted}
\title{Ridge plot weighted}
\usage{
ridge_plot_weighted(your_data, text_size = 20)
}
\arguments{
\item{your_data}{Your data frame.}

\item{text_size}{The size of the text in the plot.}
}
\value{
Bias plot for two lineages over time.
}
\description{
Given a data frame organized with two alternating cell lineages, gives ridge plots showing percent total contribution to both lineages.
}
\examples{
ridge_plot_weighted(datum[,c("T 1m", "B 1m", "T 3m" "B 3m")])
}
