% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_clone_tracker.R
\name{single_clone_tracker}
\alias{single_clone_tracker}
\title{Single Clone Tracker}
\usage{
single_clone_tracker(your_data, n_clones = 100, y_max = 1, line_size = 2,
  top_value = 0.5, text_size = 15)
}
\arguments{
\item{your_data}{Your data frame.}

\item{n_clones}{Number of clones to choose.}

\item{y_max}{Maximum value of plot.}

\item{line_size}{Size of the lines.}

\item{top_value}{Max point at which to set the gradient to red.}

\item{text_size}{Text size for the whole plot.}
}
\value{
Line plot for a chosen lineage over time.
}
\description{
Given a data frame organized from first timepoint to latest, plots top N clones over time.
}
\examples{
single_clone_tracker(datum[,c("T 1m", "T 3m" "T 5m")], y_max = 0.1, top_value = 0.05)
}

