% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hist.R
\name{stat_hist}
\alias{stat_hist}
\title{Stat histogram}
\usage{
stat_hist(your_SE, metadata_stat, group_by = NULL, text_size = 20, alpha = 0.5)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{metadata_stat}{The column of metadata to display in a histogram. Must be numeric.}

\item{group_by}{Group and color the histogram by another column of metadat. If NULL, no grouping applied}

\item{text_size}{Size of text.}

\item{alpha}{The transparency of the histograms if group_by is provided. Lower = more transparent. 1 = opaque}
}
\description{
Given a summarized experiment, gives a histogram of any piece of metadata.
}
\examples{
stat_hist(your_se = SE, metadata_stat = "GFP_percent")
}
