% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_SE.R
\name{subset_SE}
\alias{subset_SE}
\title{subset_SE}
\usage{
subset_SE(your_SE, ...)
}
\arguments{
\item{your_SE}{A SummarizedExperiment object.}

\item{...}{Arguments passed to subset_SE in the form of `X = keys` where `X` is a column from SE's colData and `keys` are entries in the colData to subset.}
}
\value{
Returns a subsetted SummarizedExperiment object.
}
\description{
Subsets an existing SummarizedExperiment object.
}
\examples{
wu_B.5month <- subset_SE(wu_subset, celltype = "B", timepoint = "6.0")
}
