% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Threshold}
\usage{
threshold(your_data, thresh = 2000)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{thresh}{Numeric.}
}
\value{
A data frame where all rows (barcodes) that did not have at least one elemnt meet the threshold have been discarded.
}
\description{
This function takes in sequence data in table form, along with a threshold,
the threshold.
}
\examples{
threshold(zh33, thresh = 2000)
}

