% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topclonescontrib.R
\name{topclones_barchart}
\alias{topclones_barchart}
\title{Top Clones Bar Chart}
\usage{
topclones_barchart(your_data, top_clones_choice, n_clones = 10,
  text_size = 15, y_limit = 100, other_color = "black",
  your_title = "Bar Chart")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{top_clones_choice}{A data frame of one column, with rownames being the barcodes. Used to get the top clones.}

\item{n_clones}{Numeric. Number of top clones from each month that should be displayed.}

\item{text_size}{Numeric. Size of text in plot.}

\item{y_limit}{Numeric. Limit of y axis in plot.}

\item{other_color}{Color to use for representation of non-top clones in the bar chart.}

\item{your_title}{Your title for the plot.}
}
\value{
Displays a bar chart (made by ggplot2) of the samples' top clones in other samples.
}
\description{
Makes a histogram of your data frame, showing the percentage contribution from
the top N clones from a chosen dataset in colors.
}
\examples{
topclones_barchart(your_data = zh33_Tlineage, top_clones_choice = zh33T1m, y_limit = 50)
}

