% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topclonescontrib.R
\name{topclonescontrib}
\alias{topclonescontrib}
\title{Top Clones Contributions}
\usage{
topclonescontrib(your_data, n_clones = 10, linesize = 2, pointsize = 3,
  your_title = "")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{n_clones}{Numeric. The number of top clones to take from columnChoice_Name.}

\item{linesize}{Numeric. Size of the lines.}

\item{pointsize}{Numeric. Size of the points.}

\item{your_title}{Character. Title for the plot.}
}
\value{
Prints a plot (using ggplot2) of the percentage contribution from each sample.
}
\description{
Makes a line plot of your data frame, plotting the percentage contribution from
the top N clones from each column as a line.
}
\examples{
topclonescontrib(your_data = zh33[,c(1:6)], n_clones = 100, your_title = "SAMPLE")
}

