% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topclonestracker.R
\name{topclonestracker}
\alias{topclonestracker}
\title{Top Clones Tracker}
\usage{
topclonestracker(your_data, top_clones_choice, months = c(1:ncol(your_data)),
  n_clones = 10, linesize = 0.5, text_size = 15, y_limit = 100,
  plot_theme = "classic")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{months}{Numeric vector. Vector corresponding to the months of each sample.}

\item{n_clones}{Numeric. Number of top clones from each month that should be displayed.}

\item{linesize}{Numeric. Thickness of the lines.}

\item{text_size}{Numeric. Size of text in plot.}

\item{y_limit}{Numeric. Limit of y axis in plot.}

\item{plot_theme}{Character. One of "BW", "classic", or "original".}

\item{your_data}{A data frame of one column, with rownames being the barcodes. Used to get the top clones.}
}
\value{
Displays a stacked area plot (made by ggplot2) of the samples' top clones.
}
\description{
Given a chosen sample, will track the top clones from that samples over the other samples included in the dataset.
}
\examples{
topclonestracker(your_data = zh33_Tlineage, top_clones_choice = zh33T1m, months = c(1,2,3,4,5,6.5, 9.5, 12), y_limit = 50)
}

