% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unilineage_bias.R
\name{unilineage_bias}
\alias{unilineage_bias}
\title{unilineage_bias}
\usage{
unilineage_bias(your_data, CT, TP, percent_thresh = 0.01, ratio_thresh = 10,
  line_months = c(1:TP), only_biased = FALSE, plot_mode = "",
  text_size = 20, line_size = 3, dot_size = 4, your_title = "",
  y_upper = 1, y_lower = 0, cellnote_display = "stars",
  by_celltype = FALSE, print_table = FALSE)
}
\arguments{
\item{your_data}{A data frame. Arranged by time point.}

\item{CT}{The number of cell types in your data frame.}

\item{TP}{The number of time points in your data frame.}

\item{percent_thresh}{The percent threshold the unilineage clones must be above.}

\item{ratio_thresh}{The ratio unilineage clones must be above to be considered biased.}

\item{line_months}{Numerical vector giving the exact months of each of the time points.}

\item{only_biased}{Logical. Whether or not to plot only the biased clones.}

\item{plot_mode}{Character. One of "heatmap", "bars", or "lines".}

\item{text_size}{Numerical size of all the text.}

\item{line_size}{Numerical size of the line thickness in "lines".}

\item{dot_size}{Numerical size of the dots in "lines".}

\item{your_title}{Character. Your title for the plot.}

\item{y_upper}{Numerical y upper limit for the "lines" and "bars" plots.}

\item{y_lower}{Numerical y lower limit for the "lines" and "bars" plots.}

\item{cellnote_display}{Character. One of "stars" or "percents" for "heatmap" plots.}

\item{by_celltype}{Logical. Whether or not to order by cell type instead of time point.}

\item{print_table}{Logical. Prints the table to be plotted rather than plotting it.}
}
\value{
Displays a heatmap, line plot, or bar chart in the current plot window.
}
\description{
Methods for displaying clones biased towards a specific lineage.
}
\examples{
unilineage_bias(your_data = zh33_byTP, CT = 5, TP = 7, your_title = "NK Bias", plot_mode = "heatmap")
}

