% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_binary_heatmap.R
\name{barcode_binary_heatmap}
\alias{barcode_binary_heatmap}
\title{Barcode Binary Heatmap}
\usage{
barcode_binary_heatmap(
  your_SE,
  plot_labels = NULL,
  threshold = 0,
  label_size = 1
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{plot_labels}{Vector of x axis labels. Defaults to colnames(your_SE).}

\item{threshold}{A chosen threshold (set as a proportion, i.e. 0.01 is 1%) that decides whether a feature entry is set to 0. Defaults to 0.}

\item{label_size}{The size of the column labels.}
}
\value{
Displays a binary heat map in the current plot window.
}
\description{
Creates a binary heatmap showing the presence of new clones from L to R in the dataset.
}
\examples{
barcode_binary_heatmap(your_SE = ZH33_SE[,1:10])
}
