% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_binary_heatmap.R
\name{barcode_binary_heatmap}
\alias{barcode_binary_heatmap}
\title{Barcode Binary Heat Map}
\usage{
barcode_binary_heatmap(
  your_data,
  names = colnames(your_data),
  your_title = "",
  label_size = 1,
  percent_threshold = 0
)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{names}{Character vetor}

\item{your_title}{The title.}

\item{label_size}{The size of the column labels.}

\item{threshold}{A chosen (percentage) threshold that decides if a barcode exists or not.}
}
\value{
Displays a binary heat map in the current plot window.
}
\description{
Creates a binary heatmap showing the presence of new clones from L to R in the dataset.
}
\examples{
barcode_binary_heatmap(your_data = zh33, percent_threshold = 0.01, label_size = 3)
BBHM(your_data = zh33)
}
