% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCheatmap.R
\name{BCheatmap}
\alias{BCheatmap}
\title{BCheatmap (Barcode Heatmap)}
\usage{
BCheatmap(your_data, names = colnames(your_data), n_clones = 10,
  your_title = "", grid = TRUE, columnLabels = 1, dendro = "none",
  star_size = 1, printtable = FALSE, table_option = "percents",
  log_transform = TRUE, log_choice = exp(1),
  distance_method = "euclidean", minkowski_power = 1,
  cellnote_option = "stars", hclust_linkage = "complete")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{names}{Names for each column of the data frame.}

\item{n_clones}{The top 'n' clones to plot.}

\item{your_title}{The title for the plot.}

\item{grid}{Logical. Include a grid or not in the heatmap.}

\item{columnLabels}{The size of the column labels.}

\item{dendro}{Whether or not to rearrange data with dendrogram.}

\item{star_size}{The numerical size of the cell note labels.}

\item{printtable}{Logical. Prints percent contribution as a table instead of plotting it.}

\item{table_option}{Character. One of "logs", "reads", or "percents".}

\item{log_transform}{Logical. Log transform data before clustering and plotting.}

\item{log_choice}{Data is log transformed with this log.}

\item{distance_method}{Character. One of "euclidean", "minkowski", "canberra", manhatan", "maximum", or "binary"}

\item{minkowski_power}{The power of the Minkowski distance (if minkowski is used).}

\item{cellnote_option}{Character. One of "stars", "reads", or "percents"}

\item{hclust_linkage}{Character. One of one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}
}
\value{
Displays a heatmap in the current plot window.
}
\description{
Creates a heatmap using the top 'n' rows from each column. More or less a wrapper for gplots' "heatmap.2" function :)
}
\examples{
BCheatmap(your_data = zh33, names = colnames(zh33), n_clones = 10,
      your_title = "First Time Point", grid = TRUE, columnLabels = 3)
BCheatmap(your_data = zh33, n_clones = 10, printtable = TRUE)
}

