% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_contribution.R
\name{clonal_contribution}
\alias{clonal_contribution}
\title{Clonal contribution plot}
\usage{
clonal_contribution(
  your_SE,
  SAMPLENAME_choice = NULL,
  clone_sequences = NULL,
  n_clones = 10,
  graph_type = "bar",
  filter_by = NULL,
  filter_selection = NULL,
  plot_over,
  plot_over_display_choices = NULL,
  keep_numeric = TRUE,
  plot_non_selected = TRUE,
  linesize = 0.2,
  text_size = 15,
  your_title = ""
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{SAMPLENAME_choice}{The identifying SAMPLENAME from which to obtain the top "n_clones" clones. If NULL, must specify clone_sequences.}

\item{n_clones}{Numeric. Number of top clones from SAMPLENAME_choice that should be displayed.}

\item{graph_type}{Choice of "bar" or "line" for how to display the clonal contribution data}

\item{filter_by}{Name of metadata column to filter by e.g. Lineage}

\item{filter_selection}{The value of the filter column to display e.g. "T" (within Lineage)}

\item{plot_over}{The column of metadata that you want to be the x-axis of the plot. e.g. Month}

\item{plot_over_display_choices}{Choice(s) from the column designated in plot_over that will be used for plotting. Defaults to all.}

\item{keep_numeric}{If plot_over is numeric, whether to space the x-axis appropriately according to the numerical values.}

\item{plot_non_selected}{Plot clones not in clones_sequences or that aren't top clones in SAMPLENAME_choice}

\item{linesize}{Numeric. Thickness of the lines.}

\item{text_size}{Numeric. Size of text in plot.}

\item{clones_sequences}{The identifying rownames within your_SE for which to plot. If NULL, must specify SAMPLENAME_choice.}
}
\value{
Displays a stacked area line or bar plot (made by ggplot2) of the samples' top clones.
}
\description{
Bar or line plot of percentage contribution from a selected choice or number of elements in the rows of the SummarizedExperiment object. Usually used for tracking a cell lineage's top clones over time.
}
\examples{
clonal_contribution(your_data = ZG66_simple_data, graph_type = "bar",  SAMPLENAME_choice = "Gr_3m", plot_over = "Timepoint", n_clones = 20)
}
