% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_diversity.R
\name{clonal_diversity}
\alias{clonal_diversity}
\title{Clonal Diversity}
\usage{
clonal_diversity(your_data, threshold = 2000, divindex = "shannon")
}
\arguments{
\item{your_data}{A data frame. Individual barcodes in rows and samples in columns.}

\item{divindex}{Character. One of "shannon", "simpson", or "invsimpson".}
}
\value{
Vector of the diversities of each column.
}
\description{
Gives the diversity for each sample (column) in your data.
}
\examples{
clonaldiversity(your_data = zh33, divindex = "shannon")
}
