% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Correlation Plot}
\usage{
cor_plot(your_data, names = colnames(your_data), thresh = 0,
  your_title = "", method_corr = "pearson", labelsizes = 1,
  plottype = "color", printtables = FALSE, no_negatives = FALSE,
  show_grid = TRUE, colorscale = "default")
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{names}{Character vector. Labels for each sample.}

\item{thresh}{Numeric. Threshold applied to your_data and scaled for each sample. When finding correlation,
only barcodes whose sum in pair of samples being compared is greater than the average of each's
scaled threshold will be used for finding correlation.}

\item{your_title}{Title for plot.}

\item{method_corr}{Character. One of "pearson", "spearman", or "kendall".}

\item{labelsizes}{Numeric. Size of the plot labels.}

\item{plottype}{Character. One of "circle", "square", "ellipse", "number", "shade", "color", or "pie".}

\item{printtables}{Logical. Whether or not to print tables of p-values, confidence intervals, and R values instead of
displaying a plot.}

\item{show_grid}{Logical. Wheter to show grid or not.}

\item{colorscale}{Character. One of "default", "rainbow", or "white_heat"}

\item{nonegatives}{Logical. Whether to make negative correlations = 0. Should almost always be TRUE.}
}
\value{
Plots correlation plot of the data frame.
}
\description{
Gives the pairwise correlation between each sample-sample pair in the data frame.
}
\examples{
cor_plot(your_data = zh33, thresh = 874, your_title = "All Samples", plottype = "ellipse")
cor_plot(your_data = zh33, thresh = 874, printtables = TRUE)
}

