% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE}
\alias{create_SE}
\title{create_SE}
\usage{
create_SE(
  your_data = NULL,
  meta_data = NULL,
  threshold = 5e-04,
  log_base = exp(1),
  scale_factor = 1e+06
)
}
\arguments{
\item{your_data}{A data frame. For cellular barcoding data, this will usually be individual barcodes in rows and samples in columns.}

\item{meta_data}{A data frame containing all meta-data. Must, at the very least, include a column called "SAMPLENAME" that contains all the colnames within the data frame passed as  `your_data` and only those colnames.}

\item{threshold}{Optional. A numeric indicating which reads to filter out (only those matrix entries that compose higher than `threshold` percentage of reads within a sample will be kept as non-zero) prior to creating the SE object.}

\item{log_base}{A numeric indicating which base to use when logging the percentages.}
}
\description{
Creates a SummarizedExperiment object from a data frame and associated meta-data.
}
\examples{
create_SE(your_data = ZH33_file, meta_data = ZH33_meta_data)
}
