% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hematoper.R
\name{hematoper}
\alias{hematoper}
\title{Hematopoietic Contribution}
\usage{
hematoper(your_data, months, n_clones = 10, scale_percent = FALSE,
  linesize = 0.5)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{months}{Numeric vector. Vector corresponding to the months of each sample.}

\item{n_clones}{Numeric. Number of top clones to view from each sample.}

\item{scale_percent}{Logical. Scale as percent of hematopoiesis or keep as number of reads.}

\item{linesize}{Numeric. Thickness of the lines.}
}
\value{
Displays a stacked area plot (made by ggplot2) of the samples' top clones.
}
\description{
Usually used for tracking a cell lineage's top clones over time. Best to
put samples in order from left to right (i.e. "1mT", "2mT", "3mT", etc.)
}
\examples{
hematoper(your_data = zh33_Tlineage, months = c(1,2,3,4,5,6.5, 9.5, 12), scale_percent = TRUE)
}

