% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_plot.R
\name{mds_plot}
\alias{mds_plot}
\title{MDS Plot}
\usage{
mds_plot(
  your_SE,
  group_by = "SAMPLENAME",
  method_dist = "bray",
  assay = "percentages",
  your_title = NULL,
  point_size = 3,
  text_size = 12
)
}
\arguments{
\item{your_SE}{A Summarized Experiment object.}

\item{group_by}{Selection from colData used to label each point. Defaults to SAMPLENAME.}

\item{method_dist}{Dissimilarity index from vegan. One of "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", or "cao".}

\item{assay}{The assay to calculate the index on}

\item{your_title}{The title for the plot.}

\item{point_size}{The size of the points.}

\item{text_size}{Numeric. Size of text in plot.}
}
\value{
Plots pairwise correlation plot for the samples in your_SE.
}
\description{
Calculates a simmilarity/dissimlarity index or metrix for each sample-sample pair and reduces the resulting dist matrix into two dimensions
}
\examples{
mds_plot(your_SE = wu_SE, your_title = "MDS plot of Bray-Curtis dissimilarities", method_dist = "bray",group_by = "celltype")
}
