% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_abundance.R
\name{rank_abundance_plot}
\alias{rank_abundance_plot}
\title{Rank Abundance}
\usage{
rank_abundance_plot(your_data, dot_size = 5, text_size = 10)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{dot_size}{Numeric. Size of the points.}

\item{text_size}{Numeric. Size of text in plot.}
}
\value{
Displays a rank-abundance plot (made by ggplot2) of the samples chosen.
}
\description{
Rank-Abundance Plot of the Barcodes for each sample chosen. Uses the cumulative sum so that all
samples eventually reach 100%.
}
\examples{
rank_abundance_plot(your_data = zh33_Tlineage[,c(1,2,4)], dot_size = 4)
}

