% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_abundance_plot.R
\name{rank_abundance_plot}
\alias{rank_abundance_plot}
\title{Rank abundance plot}
\usage{
rank_abundance_plot(
  your_SE,
  point_size = 3,
  your_title = NULL,
  scale_rank = FALSE,
  text_size = 12
)
}
\arguments{
\item{point_size}{Numeric. Size of the points.}

\item{your_title}{The title for the plot.}

\item{scale_rank}{Whether or not to scale all ranks to 1 to 100 or keep numerical integer ranks.}

\item{text_size}{Numeric. Size of text in plot.}

\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}
}
\value{
Displays a rank-abundance plot (made by ggplot2) of the samples chosen.
}
\description{
Rank abundance Plot of the Barcodes for each sample chosen. Uses the cumulative sum so that all
samples eventually reach 100%.
}
\examples{
rank_abundance_plot(your_data = wu_SE, point_size = 4)
}
