% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_plot.R
\name{ridge_plot}
\alias{ridge_plot}
\title{Ridge plot}
\usage{
ridge_plot(
  your_SE,
  selection_var,
  selection_1,
  selection_2,
  plot_by,
  weighted = F,
  text_size = 16,
  scale = 1
)
}
\arguments{
\item{your_SE}{Your SummarizedExperiment of barcode data and associated metadata}

\item{plot_by}{The column of metadata to plot by. If numeric, y axis will be in increasing order. If categorical, it will follow order of metadata}

\item{weighted}{If true, the density estimation will be weighted by the overall contribution of each barcode}

\item{text_size}{The size of the text in the plot.}

\item{scale}{Sets the overlap between ridges. Larger values give more overlap.}

\item{cell_var}{The column of metadata corresponding to cell types}

\item{cell_1}{The first cell type to be compared. Will be on the RIGHT side of the ridge plot}

\item{cell_2}{The second cell type to be compared. Will be on the LEFT side of the ridge plot}
}
\value{
Bias plot for two lineages over time.
}
\description{
Given a summarized experiment, gives ridge plots showing percent total contribution to both lineages.
}
\examples{
ridge_plot(your_se = SE, cell_var = "selection_type", selection_1 = "B", selection_2 = "T", plot_by = "Timepoint")
}
