% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_plot_diego.R
\name{ridge_plot}
\alias{ridge_plot}
\title{Ridge plot}
\usage{
ridge_plot(
  your_SE,
  split_bias_on,
  bias_1,
  bias_2,
  split_bias_over,
  bias_over = NULL,
  breaks = c(10, 2, 1, 0.5),
  text_size = 10,
  linesize = 0.4,
  ncols = 1,
  weighted = TRUE,
  scale = 1
)
}
\arguments{
\item{split_bias_on}{The column in `colData(your_SE)` from which `bias_1` and `bias_2` will be chosen}

\item{bias_1}{The factor you wish to plot on the left side of the plots.}

\item{bias_2}{The factor you wish to plot on the right side of the plots.}

\item{split_bias_over}{The column in `colData(your_SE)` that you wish to observe the bias split on.}

\item{bias_over}{Choice(s) from the column designated in `split_bias_over` that will be used for plotting. Defaults to all.}

\item{text_size}{The size of the text in the plot.}

\item{linesize}{The linewidth of the stacked bars which represent individual barcodes}

\item{ncols}{Numeric. Number of columns to plot on using plot_grid from cowplot.}

\item{weighted}{Logical. Whether ot not to use weights when running the kernel density estimator.}
}
\value{
Ridge plot of log bias for two factors over another set of factors.
}
\description{
Given a summarized experiment, gives ridge plots showing percent total contribution to both lineages.
}
\examples{
ridge_plot(your_SE = SE, split_bias_on = "Lineage", bias_1 = "B", bias_2 = "T", split_bias_over = "Month", bias_over = c(1,2,4.5,12))
}
