% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Ternary Plot}
\usage{
ternary_plot(
  your_data,
  show_arrows = TRUE,
  show_ticks = TRUE,
  density_mode = FALSE
)
}
\arguments{
\item{your_data}{A data frame. Usually individual barcodes in rows and samples in columns.}

\item{show_arrows}{Logical. Display arrows or not.}

\item{show_ticks}{Logical. Display tick marks or not.}

\item{density_mode}{Logical. Uses a kernel density estimation to view concentrations of clones.}
}
\value{
Displays a ternary plot in the current plot window.
}
\description{
Creates a ternary plot showing the bias of clones towards one of three axes.
}
\examples{
ternary_plot(your_data = zh33, dot_size = 1000, density_mode = TRUE)
}
