% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Ternary Plot}
\usage{
ternary_plot(
  your_SE,
  show_arrows = FALSE,
  show_ticks = FALSE,
  density_mode = TRUE
)
}
\arguments{
\item{your_SE}{A SummarizedExperiment.}

\item{show_arrows}{Logical. Display arrows or not.}

\item{show_ticks}{Logical. Display tick marks or not.}

\item{density_mode}{Logical. Uses a kernel density estimation to view concentrations of clones.}
}
\value{
Displays a ternary plot in the current plot window.
}
\description{
Creates a ternary plot showing the bias of clones towards one of three axes.
}
\examples{
ternary_plot(your_SE = zh33, density_mode = TRUE)
}
